Update Family_History_Normal.xml
Ben and Brett, I am doing a lot of "Data Element" work and I think it is important to label the text with labels that reinforce the presence of data elements and the structure implied by the templates used to communicate the information (when entry templates are used).  The changes I made are all to the text content labeling and the links in the corresponding machine readable data. The one nuance I would like to discuss with you both at HL7 has to do with a theory I have regarding when originalText should be used versus when the text element should be used. We can discuss this while looking at FH2prob1, the clinical statement about the mother having no known problems.  I will suggest we consider adding a FH2prob1None ID around the "no known problems" phrase in the narrative above and reference this link in the originalText of the value element.  I know I am trying to really take text linking to a new level, and there may be some initial resistance, but IMO we need to move to the next level of maturity in our representation of the data for both the humans and the machines, and I believe this type of rigor shows a way forward toward greater data element clarity.